source_code/AQI_Dashboard/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
firebase_admin==7.1.0
django-cors-headers==4.3.1
